Migration script fixes (#179)
* Fixed migration for datadirs without ancients

The script was assuming that ancients would have been migrated and was
considering the numAncients-1 to be the next block to migrate but when
numAncients is zero that's a problem.

Also remved logic for  picking up where db migration left of for the
level db since it was complicating the logic and that process takes a
few seconds, which is nothing compared with the minutes taken to migrate
the ancients.

* Ensure that we set gas limit if migrating at pre-gingerbread point

+++ op-chain-ops/cmd/celo-migrate/non-ancients.go
@@ -10,7 +10,7 @@ import (
 	"github.com/ethereum/go-ethereum/log"
 )
 
-func migrateNonAncientsDb(oldDbPath, newDbPath string, lastAncientBlock, batchSize uint64) (uint64, error) {
+func migrateNonAncientsDb(oldDbPath, newDbPath string, numAncients, batchSize uint64) (uint64, error) {
 	// First copy files from old database to new database
 	log.Info("Copy files from old database (excluding ancients)", "process", "non-ancients")
 
@@ -51,22 +51,10 @@ func migrateNonAncientsDb(oldDbPath, newDbPath string, lastAncientBlock, batchSi
 	// get the last block number
 	hash := rawdb.ReadHeadHeaderHash(newDB)
 	lastBlock := *rawdb.ReadHeaderNumber(newDB, hash)
-	lastMigratedNonAncientBlock := readLastMigratedNonAncientBlock(newDB) // returns 0 if not found
 
-	// if migration was interrupted, start from the last migrated block
-	fromBlock := max(lastAncientBlock, lastMigratedNonAncientBlock) + 1
+	log.Info("Non-Ancient Block Migration Started", "process", "non-ancients", "startBlock", numAncients, "endBlock", lastBlock, "count", lastBlock-numAncients, "lastAncientBlock", numAncients)
 
-	if fromBlock >= lastBlock {
-		log.Info("Non-Ancient Block Migration Skipped", "process", "non-ancients", "lastAncientBlock", lastAncientBlock, "endBlock", lastBlock, "lastMigratedNonAncientBlock", lastMigratedNonAncientBlock)
-		if lastMigratedNonAncientBlock != lastBlock {
-			return 0, fmt.Errorf("migration range empty but last migrated block is not the last block in the database")
-		}
-		return 0, nil
-	}
-
-	log.Info("Non-Ancient Block Migration Started", "process", "non-ancients", "startBlock", fromBlock, "endBlock", lastBlock, "count", lastBlock-fromBlock, "lastAncientBlock", lastAncientBlock, "lastMigratedNonAncientBlock", lastMigratedNonAncientBlock)
-
-	for i := fromBlock; i <= lastBlock; i += batchSize {
+	for i := numAncients; i <= lastBlock; i += batchSize {
 		numbersHash := rawdb.ReadAllHashesInRange(newDB, i, i+batchSize-1)
 
 		log.Info("Processing Block Range", "process", "non-ancients", "from", i, "to(inclusve)", i+batchSize-1, "count", len(numbersHash))
@@ -94,30 +82,27 @@ func migrateNonAncientsDb(oldDbPath, newDbPath string, lastAncientBlock, batchSi
 			batch := newDB.NewBatch()
 			rawdb.WriteBodyRLP(batch, numberHash.Hash, numberHash.Number, newBody)
 			_ = batch.Put(headerKey(numberHash.Number, numberHash.Hash), newHeader)
-			_ = writeLastMigratedNonAncientBlock(batch, numberHash.Number)
 			if err := batch.Write(); err != nil {
 				return 0, fmt.Errorf("failed to write header and body: block %d - %x: %w", numberHash.Number, numberHash.Hash, err)
 			}
 		}
 	}
 
-	toBeRemoved := rawdb.ReadAllHashesInRange(newDB, 1, lastAncientBlock)
-	log.Info("Removing frozen blocks", "process", "non-ancients", "count", len(toBeRemoved))
-	batch := newDB.NewBatch()
-	for _, numberHash := range toBeRemoved {
-		rawdb.DeleteBlockWithoutNumber(batch, numberHash.Hash, numberHash.Number)
-		rawdb.DeleteCanonicalHash(batch, numberHash.Number)
+	if numAncients > 0 {
+		toBeRemoved := rawdb.ReadAllHashesInRange(newDB, 1, numAncients)
+		log.Info("Removing frozen blocks", "process", "non-ancients", "count", len(toBeRemoved))
+		batch := newDB.NewBatch()
+		for _, numberHash := range toBeRemoved {
+			rawdb.DeleteBlockWithoutNumber(batch, numberHash.Hash, numberHash.Number)
+			rawdb.DeleteCanonicalHash(batch, numberHash.Number)
+		}
+		if err := batch.Write(); err != nil {
+			return 0, fmt.Errorf("failed to delete frozen blocks: %w", err)
+		}
+		log.Info("Removed frozen blocks, still in leveldb", "process", "non-ancients", "removedBlocks", len(toBeRemoved))
 	}
-	if err := batch.Write(); err != nil {
-		return 0, fmt.Errorf("failed to delete frozen blocks: %w", err)
-	}
+	migratedCount := lastBlock - numAncients + 1
+	log.Info("Non-Ancient Block Migration Ended", "process", "non-ancients", "migratedBlocks", migratedCount)
 
-	// if migration finished, remove the last migration number
-	if err := deleteLastMigratedNonAncientBlock(newDB); err != nil {
-		return 0, fmt.Errorf("failed to delete last migration number: %w", err)
-	}
-
-	log.Info("Non-Ancient Block Migration Ended", "process", "non-ancients", "migratedBlocks", lastBlock-fromBlock+1, "removedBlocks", len(toBeRemoved))
-
-	return lastBlock - fromBlock + 1, nil
+	return migratedCount, nil
 }
